#906 Divided flow unit tests
